+++ demo/client/src/pages/datatable.js
@@ -47,6 +47,7 @@ function initColumnVisibilityTable() {
     ],
     pageLength: 3,
     buttons: [
+      // Button to toggle column visibility.
       {
         extend: 'colvis',
         text: 'Show / hide columns',
@@ -56,36 +57,98 @@ function initColumnVisibilityTable() {
         // columns: [1,2,3,4,5],
       }
     ],
-    stateSave: true,// Save the column visibility in the browser.
+    // Save the column visibility in the browser.
+    stateSave: true,
+    /* Enable/Disable saving for various datatables elements. Delete any items you do not wish to save.
+        - data.columns.search: This option allows for the saving of the search applied to individual columns to be enabled or disabled.
+        - data.columns.visible: This option allows for the saving of the visibility of the columns to be enabled or disabled.
+        - data.length: This option allows for the saving of the page length to be enabled or disabled.
+        - data.order: This option allows for the saving of the tables column sorting to be enabled or disabled.
+        - data.paging: This option allows for the saving of the paging to be enabled or disabled.
+        - data.scroller: This option allows for the saving of the scroller position to be enabled or disabled.
+        - data.search: This option allows for the saving of the search to be enabled or disabled.
+        - data.searchBuilder: This option allows for the saving of the searchBuilder state to be enabled or disabled.
+        - data.searchPanes: This option allows for the saving of the searchPanes state to be enabled or-  disabled.
+        - data.select: This option allows for the saving of the select state to be enabled or disabled.
+    */
     stateSaveParams: (_, data) => {
-      // This option allows for the saving of the search applied to individual columns to be enabled or disabled.
       delete data.columns.search;
-
-      // This option allows for the saving of the visibility of the columns to be enabled or disabled.
       // delete data.columns.visible;
-
-      // This option allows for the saving of the page length to be enabled or disabled.
       delete data.length;
-
-      // This option allows for the saving of the tables column sorting to be enabled or disabled.
       delete data.order;
-
-      // This option allows for the saving of the paging to be enabled or disabled.
       delete data.paging;
-
-      // This option allows for the saving of the scroller position to be enabled or disabled.
       delete data.scroller;
-
-      // This option allows for the saving of the search to be enabled or disabled.
       delete data.search;
-
-      // This option allows for the saving of the searchBuilder state to be enabled or disabled. 
       delete data.searchBuilder;
-
-      // This option allows for the saving of the searchPanes state to be enabled or disabled. 
       delete data.searchPanes;
+      delete data.select;
+    }
+  });
+}
 
-      // This option allows for the saving of the select state to be enabled or disabled.
+function initColumnVisibilityWithIconButtonTable() {
+  return new components.Datatable(ref.columnVisibilityWithIconButtonTable, {
+    dom: `<'row align-items-center'<'col dataTables_pager'p><'col-auto'B>><'row'<'col-12'tr>><'row'<'col-12 dataTables_pager'p>>`,
+    columnDefs: [
+      {targets: 0, data: 'name'},
+      {targets: 1, data: 'position'},
+      {targets: 2, data: 'office'},
+      {targets: 3, data: 'age'},
+      {targets: 4, data: 'startDate'},
+      {targets: 5, data: 'salary'},
+    ],
+    pageLength: 3,
+    buttons: {
+      dom: {
+        button: {
+          // Disable the default button class (btn btn-secondary) to set the icon button (btn-icon).
+          className: null,
+        },
+        buttonLiner: {
+          // By default, a span element is added within the button, so disable it.
+          tag: null,
+        }
+      },
+      buttons: [
+        // Button to toggle column visibility.
+        {
+          extend: 'colvis',
+          columns: ':eq(1),:eq(2),:eq(3),:eq(4),:eq(5)',
+          text: '<i class="ki-solid ki-gear fs-1"></i>',
+
+          // Icon button style. Switch between them as you wish.
+          // className: 'btn btn-sm btn-icon btn-color-primary btn-active-light-primary',
+          className: 'btn btn-icon btn-color-gray-500 btn-active-color-primary',
+
+          // Display column drop-downs inside the table container. If this is not set, the column dropdowns will overflow the table container.
+          align: 'container',
+        }
+      ],
+    },
+    // Save the column visibility in the browser.
+    stateSave: true,
+    /* Enable/Disable saving for various datatables elements. Delete any items you do not wish to save.
+        - data.columns.search: This option allows for the saving of the search applied to individual columns to be enabled or disabled.
+        - data.columns.visible: This option allows for the saving of the visibility of the columns to be enabled or disabled.
+        - data.length: This option allows for the saving of the page length to be enabled or disabled.
+        - data.order: This option allows for the saving of the tables column sorting to be enabled or disabled.
+        - data.paging: This option allows for the saving of the paging to be enabled or disabled.
+        - data.scroller: This option allows for the saving of the scroller position to be enabled or disabled.
+        - data.search: This option allows for the saving of the search to be enabled or disabled.
+        - data.searchBuilder: This option allows for the saving of the searchBuilder state to be enabled or disabled.
+        - data.searchPanes: This option allows for the saving of the searchPanes state to be enabled or-  disabled.
+        - data.select: This option allows for the saving of the select state to be enabled or disabled.
+    */
+    stateSaveParams: (_, data) => {
+      delete data.columns.search;
+      // delete data.columns.visible;
+      delete data.length;
+      delete data.order;
+      delete data.paging;
+      delete data.scroller;
+      delete data.search;
+      delete data.searchBuilder;
+      delete data.searchPanes;
       delete data.select;
     }
   });
@@ -133,6 +196,7 @@ const ref = components.selectRef();
 const basicTable = initBasicTable();
 const serverSideProcessingTable = initServerSideProcessingTable();
 const columnVisibilityTable = initColumnVisibilityTable();
+const columnVisibilityWithIconButtonTable = initColumnVisibilityWithIconButtonTable();
 const disableFirstAjaxCallTable = initDisableFirstAjaxCallTable();
 
 // Initialize events, etc.